refactor code to remove the lib directory
- add services directory to store dependencies
