+++ src/services/filesystem.ts
@@ -1,9 +1,8 @@
 import fs from "fs";
-import path from "node:path";
 import {execSync} from "child_process";
 
-import {execSyncOptions, extensionName, segmentListFilename, supportedExtensions, tsInput} from "../lib/config.js";
-import {errorMsgFormatter, sexagesimalFormat} from "./formatter.js";
+import {errorMsgFormatter, sexagesimalFormat} from "../utils/formatter.js";
+import {execSyncOptions, extensionName, segmentListFilename, tsInput} from "../utils/config.js";
 
 import type {MergeOptions, RemoveVideoSegmentArguments} from "../types/index.js";
 
@@ -18,59 +17,6 @@ export const checkVideoFile = (videoFile: string) => {
             errorMsgFormatter(`${videoFile} was not found. Make sure to put the correct video filename at the top (Line 1) of ${tsInput}.`)
         );
     }
-
-    checkFileExtension(videoFile)
-}
-
-const checkFileExtension = (videoFile: string) => {
-    const extensionName = path.extname(videoFile).toLowerCase().slice(1);
-    const extensionError = supportedExtensions.indexOf(extensionName);
-
-    if (extensionError === -1) {
-        throw new Error(
-            errorMsgFormatter(`The video format ${extensionName} is not supported.
-Only the following extensions are valid:
-    ${supportedExtensions}`)
-        )
-    }
-}
-
-export const getVideoSegmentErrors = (videoSegments: string[], videoSegmentDurations: number[], baseOutputPath: string) => {
-    let possibleErrors: string[] = [];
-
-    videoSegments.forEach((file, index) => {
-        let durationInSeconds = Number(
-            execSync(
-                `ffprobe -v warning -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 "${baseOutputPath}/${file}"`
-            ).toString()
-        );
-
-        if (videoSegmentDurations[index] === undefined) {
-            throw new Error(
-                errorMsgFormatter(`Duration of video segment for index ${index} might be undefined.`)
-            )
-        }
-
-        let difference = Math.abs(videoSegmentDurations[index] - durationInSeconds).toFixed(4);
-        let isGreaterThanOne = Number(difference) > 1;
-
-        if (isGreaterThanOne) {
-            possibleErrors.push(file);
-        }
-
-        console.log(`\n[\x1b[94m${file}\x1b[0m] Duration: Computed (${sexagesimalFormat(
-                videoSegmentDurations[index]
-            )}) vs Actual (${sexagesimalFormat(
-                durationInSeconds
-            )}). Difference: ${difference} seconds.${
-                isGreaterThanOne
-                    ? "\x1b[31m Possible Error!\x1b[0m"
-                    : "\x1b[32m Result Okay!\x1b[0m"
-            }`
-        );
-    });
-
-    return possibleErrors
 }
 
 const createSegmentList = (videoSegments: string[], baseOutputPath: string) => {
@@ -153,5 +99,18 @@ Removing ${segmentListFilename}. . .`
 Video trimmer has finished. Video output should be about ${sexagesimal} long. 
 Total processing time: ${sexagesimalFormat(timeDiff / 1000)}`
     );
-
 };
+
+export const readTimestamps = () => {
+    let ts = "";
+
+    try {
+        ts += fs.readFileSync(tsInput);
+    } catch (error) {
+        throw new Error(
+            errorMsgFormatter("The file [timestamps.txt] was not found!")
+        );
+    }
+
+    return ts;
+}
